# Docs/API_SPEC.yaml
openapi: 3.0.3
info:
  title: BOSSNYUMBA Property Management API
  description: |
    Comprehensive REST API for the BOSSNYUMBA multi-tenant property management SaaS platform.
    
    ## Authentication
    All endpoints (except `/auth/login`, `/auth/register`, `/auth/forgot-password`, `/auth/demo-users`,
    and `/payments/mpesa/callback`) require a valid JWT Bearer token in the `Authorization` header.
    
    ## Multi-Tenancy
    All data is scoped to the authenticated user's tenant. The tenant context is derived from the JWT token.
    
    ## Pagination
    List endpoints support pagination via `page` and `pageSize` query parameters.
    Responses include a `pagination` object with `page`, `pageSize`, `totalItems`, `totalPages`,
    `hasNextPage`, and `hasPreviousPage` fields.
    
    ## Error Handling
    All errors follow a consistent format:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Human-readable description"
      }
    }
    ```
  version: 1.0.0
  contact:
    name: BOSSNYUMBA Engineering
    email: engineering@bossnyumba.com
  license:
    name: Proprietary
    url: https://bossnyumba.com/terms
servers:
  - url: http://localhost:4000/api/v1
    description: Local development
  - url: https://api.bossnyumba.com/api/v1
    description: Production
...
security:
  - BearerAuth: []
paths:
...
  /customers:
    get:
      tags: [Customers]
      summary: List customers
      operationId: listCustomers
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, VERIFIED, REJECTED]
      responses:
        '200':
          description: List of customers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
    post:
      tags: [Customers]
      summary: Create a customer
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /customers/{id}:
    get:
      tags: [Customers]
      summary: Get customer by ID (includes leases, current unit)
      operationId: getCustomer
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Customer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Customers]
      summary: Update customer profile
      operationId: updateCustomer
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerRequest'
      responses:
        '200':
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    delete:
      tags: [Customers]
      summary: Soft-delete customer
      operationId: deleteCustomer
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Customer deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /customers/{id}/leases:
    get:
      tags: [Customers]
      summary: Get customer leases
      operationId: getCustomerLeases
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Customer leases
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
  /customers/{id}/payments:
    get:
      tags: [Customers]
      summary: Get customer payment history
      operationId: getCustomerPayments
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Customer payments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
  /customers/{id}/kyc:
    put:
      tags: [Customers]
      summary: Update customer KYC status
      operationId: updateCustomerKyc
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [PENDING, VERIFIED, REJECTED]
      responses:
        '200':
          description: KYC status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /customers/{id}/blacklist:
    put:
      tags: [Customers]
      summary: Blacklist a customer
      operationId: blacklistCustomer
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Customer blacklisted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # ===========================================================================
  # LEASE ENDPOINTS
  # ===========================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from POST /auth/login.
        Include as: `Authorization: Bearer <token>`
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource unique identifier
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number (1-indexed)
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: Items per page
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Resource not found
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: NOT_FOUND
            message:
              type: string
              example: Resource not found
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
    MessageResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            message:
              type: string
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            type: object
        pagination:
          $ref: '#/components/schemas/Pagination'
    Pagination:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        totalItems:
          type: integer
        totalPages:
          type: integer
        hasNextPage:
          type: boolean
        hasPreviousPage:
          type: boolean

    # =========================================================================
    # Auth schemas
    # =========================================================================
...
    CreateCustomerRequest:
      type: object
      required: [firstName, lastName, email, phone]
      properties:
        type:
          type: string
          enum: [INDIVIDUAL, COMPANY]
          default: INDIVIDUAL
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        idNumber:
          type: string
        idType:
          type: string
          enum: [NATIONAL_ID, PASSPORT, DRIVERS_LICENSE]
        companyName:
          type: string
        companyRegNumber:
          type: string
        preferences:
          type: object
    UpdateCustomerRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        preferences:
          type: object

    # =========================================================================
    # Lease schemas
    # =========================================================================
...